Unit Test File
Code for P3.6 Unit testing. This code creates a test file based off of scan data and exports it to its own custom directory, then checks if the file exists. This will print the results of the unit test to the terminal. This code also implements a change to the way that JSONs are called, the JSONs were getting a little messy in the root folder, so to clean things up I placed them all in a folder and changed the reference of the JSON file in the code to represent this change.