Fixed Arduino Connection Issue
Fixed issue with C++ program that caused it not to properly initialize a connection to the Ardiuno serial monitor. Previously you had to open the arduino ide and send a message in the serial monitor for the C++ program to be able to send a message, now the connection is initialized properly.